fix: fix stuck at loading when activity recreation happened on Feed Browse and Category Inspiration page
checking savedInstanceState null will only cover case when configuration changes happened, but when activity recreation happened, savedInstanceState might not be null and it might cause Feed Browse and Category Inspiration to not fetch the data again. To cover the activity recreation case, we can use `Activity#lastNonConfigurationInstance` instead. If it is null, most likely we need to fetch the data