`feat`: add product card compact to customer mid gradle because error: [:dynamic-features:base:df_base, :dynamic-features:tokopedianow:df_tokopedianow] all package the same library [:features:discovery:productcard_compact;Capability: group='android-tokopedia-core.features.discovery', name='productcard_compact']
Multiple APKs packaging the same library can cause runtime errors. Placing each of the above libraries in its own dynamic feature and adding that feature as a dependency of modules requiring it will resolve this issue. Libraries that are always used together can be combined into a single feature module to be imported by their dependents. If a library is required by all feature modules it can be added to the base module instead.